uac-platform/controller/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
redis==5.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.3
httpx==0.26.0
PyYAML==6.0.1
